+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [UNRELEASED]
+## v.0.5.0 (30 June 2022)
 - Adds capabilities: "TOW" and "AHV" for tugboat/towing equipment and anchor-handling vessels
 - Adds a tow-to-port strategy that is activated for repairs with the "TOW" capability in the servicing
 - Adds a `Port` class to handle the tow-to-port class and tugboat-based service requests
@@ -10,6 +10,11 @@
 - Adds a `location` flag to the events logging infrastructure and implements its usage across the simulation architecture
 - Creates the metric `Metrics.number_of_tows` to track the number of tows and provides breakdowns as needed
 - Creates the metric `Metrics.vessel_crew_hours_at_sea` to track the number of vessel or crew hours at sea
+- Creates the metric `Metrics.port_fees` to calculate any port fees associated with a project
+- Creates the metric `Metrics.opex` to calculate a project's operational expenditures
+- Creates the metric `Metrics.NPV` to calculate a project's net present value
+- Modifies `Metrics.project_fixed_costs` to have more time resolutions to align with the OpEx calculation options
+- Fixes some results formatting inconsistencies in the `Metrics` class
 
 ## 0.4.1 (2022-March-8)
 - Adds code diagrams to demonstrate how the various components connect